Don't add always_inline to most functions
This isn't necessary and defeats the purpose of
calls in many cases. Many of the large transcendental
functions are good candidates to call. This should
rely on the inliner heuristics to be inlined. Today
one of the first things done after linking is to annotate
every function with alwaysinline anyway, so there's no
reason to explicitly mark every library function this way.

Also rename the macro to be clear that it is for always
inlining. Only keep it on a few functions where it may be
useful because they will usually emit a few instructions
which are likely to codegen to one. The workitem functions
are good candidates because in practice they are always
called with constants and most of the body is deleted
(although the inliner looks for things like this as ell).