Fix half median3 for pre-GFX9
Use an expansion of median when v_med3_f16
isn't available.

I think this is correct for sNaN behavior,
at least when the definition of the llvm
intrinsics for minnum/maxnum are fixed to
follow the libm semantics.

This function doesn't seem to be
well specified in the extension
description.

Change-Id: I08d45a0fe2cea4a31c631db403c3fbed90a4d25b