cmake: Fix for parallel build with ninja
When building the tests in a parallel build,
the various .bc files usually did not yet exist.
It seems this happened to work usually with the
make build, but was consistently breaking for me
with ninja.

Because the built libraries are linked directly by
clang with -mlink-bitcode-file, these are not link time
dependencies as cmake currently believes.

Introduce an intermediate custom target to act as a
barrier to ensure the libraries are fully built before
trying to compile the source files for the tests.

Change-Id: I06b9b02b191f1246b9b2eed14a5ba4af4934456e