eliminate the indirect dependency DEPS_TGT from opencl_bc_lib
The macro opencl_bc_lib collects a list of files that are to be linked
into the target library in a phony target DEPS_TGT. When this phony
target is listed as a dependency for the target library, the resulting
dependency edge works only one way: the rule for the library ensures
that the dependencies are built first, but does not ensure that the
target library is rebuilt if the dependencies are rebuilt. The latter
part is ensured only if the dependency itself is "an executable or
library", as described in the CMake manual:

https://cmake.org/cmake/help/v3.5/command/add_custom_command.html

For example, with such a phony target, the resulting Ninja rule
contains an order-only dependency:

https://ninja-build.org/manual.html#ref_dependencies

This change replaces the phony target with a list of dependencies,
so that real file-level dependency edges are created by CMake.

Change-Id: I5fa3f8b1570041104f2e9d1cf270d25dff4b7be1