remove __ockl_as_feature_t; prefer uint for size of input data
__ockl_as_feature_t was introduced to track whether a stream supports
asynchronous operation or not. But this is equivalent to whether the
doorbell signal handle is non-zero. Hence eliminating this field.

The API call __ockl_as_write_block() expects an argument indicating
the size in bytes of the data being written to the stream. This should
not be size_t since the host and device may not use the same width for
size_t. Hence replacing that type with uint, which matches the
register size and should be sufficient for the strings that we expect
in actual usage. Larger sizes can be added subsequently.

Change-Id: I4ecbbff828b5d8886685424715fd2fa14c7314b1