hostcall: use asm volatile barrier for convergent functions
One of the test cases for hostcall messages contains the
following pattern

__ockl_hostcall_message_append_args() {
  ...
  __ockl_hostcall_preview(...);
}

__ockl_hostcall_message_append_bytes() {
  do {
    ...
    __ockl_hostcall_preview(...);
  } while (wfall() == 0);
}

kernel_foo() {
  ...
  __ockl_hostcall_message_append_args();
  ...
  __ockl_hostcall_message_append_bytes();
}

append_bytes() writes out one string per workitems. The do-while()
loop may diverge since different work items may have strings of
different lengths. Hostcall uses mbcnt followed by readfirstlane to
determine the first active lane. Since mbcnt has no side-effects, this
whole pattern can be (incorrectly) hoisted out of the loop. The use of
wfall() is supposed to prevent this hoisting.

But in this particular kernel, the compiler still manages to common
out the readfirstlane from append_bytes() due to the presence of a
previous readfirstlane from append_args.

The only way to prevent this is to create an "asm volatile" barrier
between mbcnt and readfirstlane. This pattern is now invoked only once
at the entry to hostcall, and the resulting pair of values are passed
to all the helper functions.

Change-Id: I7df07620cef4701ddfaad2a1570f945a69fee2c6