hostcall: replace index_size with index_mask
The hostcall buffer uses tagged pointers, where the index mask in the
lower few bits is used to index into the array of packets. The
index_size in the hostcall buffer is used to compute the mask for the
index. This change replaces index_size with the actual mask itself,
which improves the readability of the code and might also save a
couple of instructions. The index_size was 32-bit while the new mask
is 64-bit, but that's a minor cost to pay in the hostcall buffer.

Change-Id: I705f2040f8b0de31317b55276b184df77e714254