Record list of all bitcode targets in a global property
This is so the comgr build can query a list of all the bitcode
libraries when both projects are build as part of
LLVM_EXTERNAL_PROJECTS. After spending > 1 week of searching for a way
to pass the list of targets to other subproject builds, this is the
least bad option I've been able to come up with.

CMake does not provide a way to get a list of all targets produced by
a project(), or even globally. Generally downstream targets use
find_package, and are then able to directly use imported targets by
name. In this case we do not want to refer to a hardcoded, specific
targets, since the device library set is subject to frequent changes
whenever a new subtarget is added.

Other LLVM subprojects seem to use this technique. AddLLVM appends
libraries to LLVM_LIBS, and similarly MLIR defines a MLIR_ALL_LIBS
property.

The current variable AMDGCN_TARGETS_LIB_LIST method works in a way
that's only usable if this is directly built as a subdirectory of
another project, and requires a very strange looking variable return
pattern. This should eventually be removed, but this can be done after
the corresponding changes are made in the comgr build.

The related AMD_DEVICE_LIBS_TARGETS should also probably be
renamed. The general convention seems to be to call this
PROJECT_NAME_EXPORTED_TARGETS.

Change-Id: Ic5d2204a65b770349625f134e96cda9a2eb6dd04