Add constant folding tests for lgamma_r
The compiler should be able to constant fold the majority of
functions, and this also provides a way to test the edge cases of the
implementation assuming the constant folding is correct.

I wasn't sure how to run llvm-lit from outside the llvm directory, so
this uses cmake to run clang and FileCheck, and can be run with ctest.

Change-Id: I9e2f6590fbf754666b9e4ca30e7e8cafb9474168