Update build documentation
This for some reason was directing you to build the host components
with the built clang, which is unnecessary and just adds extra setup
complexity (it was also incorrect since it wasn't setting CXX and
there are no C components). Document the preferred cmake mechanism for
prioritizing a specific clang build. Also document that this can be
built as an llvm external subproject.

Stop referring to "AMD" llvm, as there are at least 4 different AMD
llvm forks on github, so it wasn't really specific enough. Also
mention the upstream llvm-project build should also always work.

Change-Id: I3dca346ee846bf4aecfb0bebbbf124dea4ae6be5