Eliminate irif wrappers around llvm.ctlz
__builtin_clz has a bizarre design where the behavior for 0 is target
defined in clang, to set the second operand to whether a 0 result is
defined. InstCombine handles folding checks for 0 around these, so
just rely on that to get the defined 0 behavior.

Produces identical IR after instcombine for all the types. The AMDGPU
instruction really returns -1 for 0, so we end up splitting this out
again later anyway.

Change-Id: I7a7d867275f9d098429837dc1f3bf24811713eac