Use new __builtin_elementwise_* functions instead of IR wrappers
These were added last year and work with any scalar or vector type (so
we don't really need to have the different _type suffixed macros). Not
all the builtins are covered. Currently roundeven does exist, which I
believe is equivalent to round and rint. I'm confused by having 3
names for the same operation, so leave the most likely to break for a
later change.

I have a patch out to add copysign and canonicalize. fma will be a bit
more work because the code was written only for 2 operands and has
some promotion rules to consider.

Change-Id: I8acf49ba718f32dd9a97380c78c5f95029b5d795