Replace single class test of infinities with compare equal.
Regex and clang-format, plus fix up the few places that needed new
parentheses.

Class is fancier than necessary and has stricter requirements than a
regular compare. The backend already optimizes these down to a regular
compare. In the case of f16, it also brings with it a subtarget
requirement as long as we're going through __builtin_amdgcn_classh.

Change-Id: I3dfa805608d7c987f765b7d464063e3fda63cb63