Avoid use of class builtin in cbrt
This will avoid the need for REQUIRES_16BIT_INSTS. InstCombine can now
put these back together into a class call. If you look at the IR of
the raw compiled library, it looks worse since the compare to 0 did
not fold into the class call. This is because this also fixes a buglet
where the test for 0 didn't consider DAZ. After denormal-fp-math is
overridden in the final linked program, the IR is equivalent (just
with an inverted condition+select).

Change-Id: I44b35bfb580580535cef8f1fc23c49ee3fb4b595