Use generic exp2/log2 in unsafe pow path
This was just inlining the denormal handling for log2/exp2. The scale
factor is different, but I don't think it really matters as long as
it's lifted out of the denormal range. log2/exp2 use 2^64 as the scale
factor, and this used 2^24 (I believe this could have been 2^23 to
begin with for the minimum scale). I see a 4 byte improvement in the
conformance test but I'm not 100% sure where it's coming from. Gets
the same max_ulp on bruteforce pow rlx.

Change-Id: Ifa14c50e1bec7e1e9faae4c8b86285aaab3bcb17