Adjust rsqrt scaling code
Rewrite it so folding out the multiply is a canonicalization dropping
optimization, which is a bit easier to deal with. It's legal to drop
canonicalizations without strictfp, so the multiply can be dropped
without consideration of the context. Previously we would have had to
consider the context and introduce a new multiply to preserve signed
zero handling which could later be folded into canonicalizing
contexts, which is more steps.

Some of the other scale cases are already written in a canonicalize
dropping way, e.g. cbrt and rcbrt.

Also saves a materialize of a constant pre-gfx10 and always save 4
bytes of code size.

Change-Id: I45cc1c6c4a514cc4f64d59ec60c46ba91ebe955c